offline_agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
PyYAML>=6.0.1
aiofiles>=23.2.1

# Optional desktop automation (Mode 3)
# pyautogui>=0.9.54
# Pillow>=10.0.0
# pygetwindow>=0.0.9
# pytesseract>=0.3.10
